Fixes getPortAddresses being space separated value
Addresses on LSP are a space separated value of mac and ip addresses.
This was fixed for setting the address on an LSP in:
https://github.com/ovn-org/ovn-kubernetes/commit/c7cad83baf4fd756ee7bb9a018453f8a20ef9295

However, the logic was never fixed in getPortAddresses to split the
string into mac and ip values. This leads to errors like:
failed to parse logical switch port "openshift-multus_network-metrics-daemon-qzmg7"
MAC "0a:58:0a:80:02:04 10.128.2.4": address 0a:58:0a:80:02:04 10.128.2.4: invalid MAC address

Signed-off-by: Tim Rozet <trozet@redhat.com>