libovsdb: Fill in uuid of matched row
When matching for rows is done via a predicate function, there is
a flag called BulkOp that means that up to one row should be found.
In such cases, the changes proposed here will fill in the uuid of row.

This saves caller from having to grab the uuid from the results
slice, normally done via the DoAfter attribute, like in the example below:

 1 loadBalancerGroup := nbdb.LoadBalancerGroup{ Name: types.ClusterLBGroupName, }
 2 loadBalancerGroupRes := []nbdb.LoadBalancerGroup{}
 3 opModels := []libovsdbops.OperationModel{
 4    {
 5      Model:          &loadBalancerGroup,
 6      ModelPredicate: func(lbg *nbdb.LoadBalancerGroup) bool { return lbg.Name == types.ClusterLBGroupName },
...
 7      ExistingResult: &loadBalancerGroupRes,
 8      DoAfter:        func() {
 9                          if len(loadBalancerGroupRes) > 0 {
10                              loadBalancerGroup.UUID = loadBalancerGroupRes[0].UUID
11                      }
12    },

Note that OperationModel.BulkOp is omitted because that is False by default. In this case,
lines 7 and below would no longer be needed because 'loadBalancerGroup.UUID' will already
have the wanted value populated.

Signed-off-by: Flavio Fernandes <flaviof@redhat.com>