Check if pod exists before re-adding SNAT
In order to make egressIPs and externalgws compatible,
we re-add the SNAT to nodeIP when we delete the pod
if disableSNATMultipleGWs is true. While doing this,
we need to check if the pod exists or not, because
delLogicalPort is called first before
deletePodEgressIPAssignment. This will leave stale
SNATs behind.

Also enable the CI job pipeline for disableSNAT=true
(From 4.10 OCP this is the default)

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>