Multiple ExGW: fixes unique pod exgw cache key
When pods are added to the cache as exgws for a namespace, only the
pod's name is used as the key. This breaks a scenario where 2 pods with
the same name are serving as exgws for the same namespace. Consider this
example:

1. app pod is created in ns foo
2. exgwAPod is created in ns exgw1 (172.0.1.1), serving ns foo
3. exgwAPod is created in ns exgw2 (172.0.1.2), serving ns foo

In the above example, the app pod will only have one ECMP route for
172.0.1.2, because the cache is keyed only on pod name.

Signed-off-by: Tim Rozet <trozet@redhat.com>