Set l3GWConfig.mode correctly
The "mode" field in the l3gw annotation
actually denotes if there is a gateway
or not, and it means "have we created
an interface that we share with the node".
The value for this is always "shared" for
both local and shared modes and this is
confusing. Let's just update this with
the right mode.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>