Fix cache synchronization on restart
There are two caches used for egress IP: the allocator cache, tracking
which egress IP has been assigned to which node, and the podAssignment
cache, tracking what set up has been performed on a per-pod basis. The
logic in reconcileEgressIP only warmed up the cache when new assignments
were actually being performed, overlooking the fact that we need to add
everything that is already existing and which won't be re-assigned to
the cache as well.

Moreover, when running in a cloud environment we also never cleared the
`statusToRemove` from the allocator cache before possibly trying to
re-assign new IPs. This fixes that too

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>