Pull up switch names from cache
While creating ACLs, we are using the
nodeNames as the switch names on which
the ACLs need to be created on. This
will break on hybrid-overlay mode
since not all nodes have the OVN
switches. Let's use the libovsdb cache.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
(cherry picked from commit 649883e42ea31db907f013736eb8ca8cdcfaa94a)