Fix ICNI setup/deletion for egress IP matching pods
Pre-this-patch-series: we always deleted the external GW setup for each
pod matching an egress IP and proceeded to wiring it up for egress IP.
That can happen around 20 times in a short time span since pod updates
are atomic and many controllers can change many fields on the pod when
it's created. With this series of patches: we now have something which
tracks what has been performed for each pod and do not perform repeating
work. This is also needed since we have several different handlers which
react to changes and can perform the egress IP setup, WatchEgressIP,
WatchEgressNodes, etc.

We need to be sure that the external GW setup created in addLogicalPort
has actually been performed once the egress IP code is supposed to
delete that. This was never the case before because it wasn't required
given that we always re-did the same work for every pod update and were
bound to "finish last" at some point.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>
(cherry picked from commit 9353e949c42ea6db4da7be824fd3043af18738f6)