Change watch handler initialization for EgressIP
WatchEgressIPNamespaces and WatchEgressIPPods only use the informer
cache to retrieve the egress IPs when determining if namespace/pods
match. It is thus better if we initialize them first and allow
WatchEgressNodes / WatchEgressIP to initialize after. Those handlers
might change the assignments of the existing objects. If we do the
inverse and start WatchEgressIPNamespaces / WatchEgressIPPod last, we
risk performing a bunch of modifications on the EgressIP objects when
we restart and then have these handlers act on stale data when they
sync.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>