Ensures reconciliation of duplicate load balancers in OVN
Based on scale results we see there may be instances of duplicate load
balancers in OVN, meaning two or more load balancers with the same name.
In OVN-K this name is unique, so there should never be more than one.
The cause of this issue is unknown, but this commit provides a
workaround.

This is an alternative solution to getting rid of the cache as proposed
in: https://github.com/ovn-org/ovn-kubernetes/pull/2757

The OVNK cache is important because LBs in OVN are not keyed based on
name. Therefore when looking in OVN for load balancers (which happens
often) we have to poll through all of them to search for the matching
name.

Changes-Include:
  1. rework the OVN LB cache in order to index by name. In OVNK names of
     load balancers are unique, so it makes sense to index by the thing
     we care about. It also elminates some of the searching done in the
     cache. Additionally, we allow the UUID field to be UUIDs, to track
     more than one entry in OVSDB corresponding to this load balancer.
  2. Add a call back hook into libovsdb for ovsdb cache updates to load
     balancer add operations. This hook is used to add/update the OVNK
     cache whenever a load balancer is created in OVN.
  3. Add a check during EnsureLBs to detect if a load balancer has more
     than 1 UUID. If it does, delete all of the current UUIDs and
     create a new one.

Signed-off-by: Tim Rozet <trozet@redhat.com>