Fix pod-creation-retry
This commit fixes addRetryPods to correctly
construct the retryPods cache. Currently
since we assign pointer as value in
retryPods cache in a loop, the values
get overwritten and this causes a nasty
bug in the retry pod logic. Some pods
get retried while others don't.

This PR also fixes addLogicalPort to
fail hard if the node.l3gateway annotation
is not set in disableSNATGWMode.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>