Fix update of CloudPrivateIPConfig
`executeCloudPrivateIPConfigChange` computes the diff between what needs
to be assigned and removed, and passes the operations to `executeCloudPrivateIPConfigOps`.
However, the `ops` structure used isn't a pointer based structure,
meaning that for update operations: the modifications made for removals
didn't actually modify the same struct, leading to update operations
being treated as add operations and failing later on in `executeCloudPrivateIPConfigOps`.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>