Ensure delete before add for egress IPs
On a cloud: we wait for the status on the CloudPrivateIPConfig(s) we
create to be set before executing anything in the OVN DB. So, we need
to make sure that we delete and then add, mainly because if EIP1 is
added to nodeX and then EIP2 is removed from nodeX, we might remove
the setup made for EIP1. The add/delete ordering of events is not
guaranteed on the cloud where we depend on other controllers to
execute the work for us however. By comparing the spec to the status
and applying the following truth table we can ensure that order
of events.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>