Remove event which will be spurious for cloud egress IP
On public clouds we will re-process assignments for the same IP
multiple times due to the nature of syncing each individual
CloudPrivateIPConfig one at a time. This means that we are
expected to end up in this situation multiple times per sync. Ex:
Say we an EgressIP is created with IP1, IP2, IP3. We begin by
assigning them all the first round. Next we get the
CloudPrivateIPConfig confirming the addition of IP1, leading us
to re-assign IP2, IP3, but since we've already assigned them
we'll end up here. This is not an error. What would be an error
is if the user created EIP1 with IP1 and a second EIP2 with IP1,
then we'll end up here too and that would be a "user error".

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>