Fixes finding default gateway for configured GW interface
The code for determining the default gateway was not taking into
consideration if the gateway interfaces was provided via config and
would return any default gateway found on the host. This is an issue
when a system has more than 1 default gateway, as well as if a user
provided an interface other than the real default gateway interface.

If the gateway interface is provided via config, the code now scopes
default routes found to only that interface.

Signed-off-by: Tim Rozet <trozet@redhat.com>