Fixes handling errors for getting IPs for pods
When network policy would go to add peer pods via the handler, it would
commonly fail on initial add because it would fail to get pod IPs, due
to racing with the pod handler for the pod add event. This is OK,
because eventually the pod will get an update event once the ovn
annotation is added and the network policy handler will add it. However,
this would cause an error log to be spammed, especially with a lot of
pods and network policies in place.

This commit ignores logging errors on policy add/update events where the
pod has no IP either in the OVN annotation or in the KAPI pod IP fields.

Additionally, changes GetAllPodIPs function to be more readable and
eliminates a check in namespace.go for legacy PodIP field.

Signed-off-by: Tim Rozet <trozet@redhat.com>