Fix incorrectly logged error when searching for ovnkube master pod
Before registering OVN northd metrics, we check to see if ovnkube
master is co-located on the same node and repeatedly check if it
is present because ovnkube node may start before ovnkube master.
This repeated checking only causes one error mode to occur which
is a timeout error and this is incorrectly logged as an error
confusing the user.
The only failure path to occur within 'checkPodRunsOnGivenNode' can
be seen with increased log level.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>