Fixes race for namespace logging level update
Updating the ACL logging by checking the namespace was added to the
network policy handler as part of:
https://github.com/ovn-org/ovn-kubernetes/pull/2809

However, the policy was not added to the namespace before updating the
logging levels. Therefore the current policy would not be processed
during logging level updates in this path.

Reported-by: Andrew Stoycos <astoycos@redhat.com>

Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 8e2b6437840884a69e780452269d700df64bc651)