Fixes delete retry on network policy recreation
This fixes a case where:
  1. A network policy foo/bar is deleted, but fails
  2. foo/bar with delete is added to retry cache
  3. foo/bar is recreated in kapi
  4. foo/bar gets added in network policy handler

In the above scenario we would not clean up the old network policy, and
just add the new one. If that was successful we would miss cleaning up
the old policy.

Additionally, Update now handles if there was an oldPolicy in a retry
entry and correctly removes it, otherwise tries to delete the old policy
in the update event.

Signed-off-by: Tim Rozet <trozet@redhat.com>