Fix ovnNBLSPDel func
Let's rename ovnNBLSPDel to delLSPOps since this
function no longer actually deletes anything and is
only returning the ops.

Also change the logic a bit:
1) Validate LSP first. Get the LSP from ovsdb cache.
If the LSP doesn't exist, nothing to do, return nil.

2) Get the UUID of the logical switch from logical switch manager.
If this fails we know the switch must exist, because the LSP exists.
Check OVN cache. If the switch is now gone, then we know the LSP
must also be gone, return nil.

Finally if Switch was found and LSP was validated, send delete op.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>