Using "update" instead of "mutate" when updating GatewayChassis of LRP.
To sync the logical router port, we used to delete and add it back in
the same trasanction with ovn-nbctl, making sure no existing dirty data
are left. However, now using libovsdb, we are using mutate to set the
gateway_chassis, and if the LRP has an old gateway chassis set already,
this mutation would add an extra gateway chassis instead of replacing
the old one. This can happen when the chassisID annotation changes,
resulting multiple gateway chassis set for the LRP, leading to
unexpected behavior, such as BFD being enabled.

Signed-off-by: Han Zhou <hzhou@ovn.org>