gateway/shared: snat hairpin traffic to other node addresses
There is a very special class of traffic, where a load balancer backend
is a different address on the node. This works for pod -> service ->
backend traffic, since it goes out via the management port.

This does not work for host -> pod -> host case though. So, all traffic
that matches this.

This includes a bit of glue to re-sync flows when node IPs change.

Signed-off-by: Casey Callendrello <cdc@redhat.com>