Metrics: count CNI requests and CNI requests with errors
Currently, we expose a metric that records the duration of a
CNI request and if an error occured, a label is set.
However, it does not allow us to understand the
total amount of CNI request errors. This patch adds this
information and also the total amount of requests.

This will aid creating service level indicators for
OVN-Kubernetes.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>