Fixes node retry test and updating nodes next hop
Changes the test to actually trigger an update event by updating the
node annotation with a new nexthop for the node. This revealed a bug in
the gateway update logic where we were using the nexthop as a predicate
for deciding on updating or adding a new route. Really route updates
should be based on IP prefix and not on destination.

Signed-off-by: Tim Rozet <trozet@redhat.com>