make WatchFactory addHandler retry processExisting when needed
Move the retry logic when performing sync (aka processExisting) to a
common place in the code. With that, watchers are expected to have
an idem-potent implementation of sync.

Since not all implementations are idem-potent now, this commit
added TODOs to mark those places in the code for future work.
Until then, some sync functions will log the failure but not
return from it.

Reported-at: https://issues.redhat.com/browse/SDN-2837
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>