Multicast E2E - increase server pod timeout
The test was only running the servers for 30 seconds, and then checking
the logs for 60 seconds max to see the servers received data. This would
potentially cause the servers to go to completed state, which means ovnk
will destroy all networking for those pods. In order to debug the
failure correctly, we need the pods to exist when the databases are
dumped.

Increasing the timer is safe, because on test cleanup it will destroy
the pods anyway.

Signed-off-by: Tim Rozet <trozet@redhat.com>