Fix EgressQoS pod sync panic
When using the namespace's address_set for the
EgressQoS rule we should return an empty map rather
than nil, otherwise when we try to load from it
in the pod sync path we get a nil pointer dereference.

Using an empty map is fine since we don't want to
change anything in the namespace's address_set here.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>