Add nosprintfhostport linter config
This enables the [no-sprintf-host-port](https://github.com/stbenjam/no-sprintf-host-port) linter to prevent URL
constructions that are incompatible with IPv6.

The Go linter no-sprintf-host-port checks that sprintf is not used to
construct a host:port combination in a URL. A frequent pattern is for a
developer to construct a URL like this:

```go
fmt.Sprintf("http://%s:%d/foo", host, port)
```

However, if "host" is an IPv6 address like `2001:4860:4860::8888`, the
URL constructed will be invalid. IPv6 addresses must be bracketed, like this:

```
http://[2001:4860:4860::8888]:9443
```

Signed-off-by: Stephen Benjamin <stephen@redhat.com>