pods: don't look up LSP twice
addLogicalPort() looks up the LSP early and then looks it up again
if the pod had no IP/MAC annotations. This is kinda pointless and
just costs time and lock contention in libovsdb since if the port
was found early we can just use that one to get the MAC/IPs.

Since the only other user of GetPortAddresses() was the hybrid
overlay code move the port lookup into the hybrid overlay code
and just keep the address extraction code common.

Signed-off-by: Dan Williams <dcbw@redhat.com>