Fix egressIP object deletion if the node is deleted first
There are a few instances in egressIP code where we shouldn't
fail hard if objects are not found, example if a node gets
deleted before egressIP watcher takes action. In this case,
we can log a warning saying objects couldn't be deleted and
move on since its technically not an error if the object was
deleted by some other code path.

In oder to ensure we don't leave behind stale reroute policy
entries, we also add the logic of removing policies that have
nexthop == joinIP of the node that's getting deleted into the
gateway cleanup code.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>