Update logging for specific policy when creating it
All the namespace policies ACLs were being updated when adding a single
policy to make sure that the ACL logging levels were current on the
policies in case it changed while the policy was being created. This is
unnecessary and has performance issues.

When creating a policy, only update the ACL logging of that policy, and
only if it changed from the value we created it with. The other
namespace policies will be handled through the namespace update handler.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>