Partial Revert of onModelUpdatesAll
Reverts part of https://github.com/ovn-org/ovn-kubernetes/pull/2800

onModelUpdatesAll was not really updating all...only things that had non
default values. Therefore if there was an update operation using
onModelUpdatesAll, where the desired update was to set a column back to
its default state, it would be ignored. This true for specific types
like maps, slices and pointers.

This commit reverts the behavior to its previous state before
onModelUpdatesAll was added. Additionally for those models which used
the same behavior of only updating default previously, onModelUpdatesAll
is renamed to onModelUpdatesAllNonDefault.

In the future, someone may want to implement a real onModelUpdatesAll
function which would dynamically figure out all of the fields of a model
to update and return that list. For now the code is commented with TODOs
and reverted back to its previous state.

Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 5d84deb8132459f578525e9556bae414ff7f2c0d)