Remove metric ovs_vswitchd_dp_if due to high number of time series
This metric produces a time series for each interface and this
becomes problematic when the number of nodes in a cluster scales.

For example, with a 50 node cluster and 200 pods per node, we are
looking at 10000 time series.

In order to protect the prometheus database, I want to remove this.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>