services controller, tests: stale node load balancers
This commit adds a new unit test that shows the node load balancers are
not removed upon node removal.

If the node is re-added (as part of any remediation procedure), the
newly created node logical switches will *not* point to the existing
load balancers, since the ovn-k master assures the load-balancers are
not re-created. The existing node logical switches will only be mutated
to point at these load balancers upon load balancer creation.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>