ovnkube-node: Full mode management port
Currently in the full mode ovnkube uses OVS internal port as management
interface. However in some use cases it is preferable to use real
devices, like SRIOV VFs, as management ports to utilize offloading, for
ex..
When ovnkube-node-mgmt-port-netdev parameter is specified then
corresponding representor is discovered and both netdevice & representor
are configured to serve as management port. Since representor resides on
the same host, it is renamed to "ovn-k8s-mp0_0".
If ovnkube-node-mgmt-port-netdev is not specified, then OVS internal
port configured as it is now.

Allign existing and add new unit tests.

Signed-off-by: Dmytro Linkin <dlinkin@nvidia.com>