Conditionally request pod retries during node add
We were always requesting pods on a node to be retried immediately, even
though they would end up failing if disable snat gw was true and the
node did not yet have an l3 gateway config set. If this case is true,
the pod add will fail, so there is no point requesting them to be added.

Signed-off-by: Tim Rozet <trozet@redhat.com>