network-controller-manager: Manage the nodes which belong to local zone.
This patch adds the concept of zones where a zone can have one or more
nodes grouped in it.  To start with this patch only supports one global
zone (all the nodes belong to the global zone) and upcoming patches will
add the support for multiple zones.

Each zone requires
  - OVN Northbound and Southbound database cluster
    (either a single node or a raft cluster)
  - ovnkube network-controller-manager which connects to its local zone OVN db
    cluster and creates the OVN logical resources in the Northbound
    database. network-controller-manager can be run on a single node or
    multiple nodes with active/passive HA (etcd leader election).
  - ovn-northd which connects to the local zone OVN dbs.
  - ovnkube-node and ovn-controller on each node of the zone.

network-controller-manager creates the node logical switch and other Northbound
resources for a node only if the node belongs to its zone.  Nodes with other zones
are ignored.  It also creates pod resources only if it is scheduled
in the local zone node.

network-controller-manager gets its zone name from the NB_Global.name of
OVN Northbound databse. If the name is empty then the default value "global" is used.

Signed-off-by: Numan Siddique <numans@ovn.org>