Add nil pointer check to destroyNetworkPolicy
If for any reason destroyNetworkPolicy is called with a nil pointer,
make sure to report such an error instead of causing a nil pointer
dereference exception. For example, this might happen if
deleteNetworkPolicy is fed a nil pointer network policy and if no entry
is found inside nsInfo for the policy.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>