Improve memory efficiency when checking stale VF representor
checkForStaleOVSRepresentorInterfaces was found to run
in non-SRIOV deployment which causes memory spike every
minute with extra 512kB memory allocated in the heap.
This could be a problem when running ovnk in a small
form factor deployment.

The following is observed via pprof:
512.14kB 14.17% 57.50%   512.14kB 14.17%  github.com/ovn-org/ovn-kubernetes/go-controller/pkg/node.checkForStaleOVSRepresentorInterfaces

this commit filters "ovs-vsctl find" output with "vf-netdev-name"
external-ids so it takes much less memory in non-SRIOV case.

Signed-off-by: Zenghui Shi <zshi@redhat.com>