services controller, tests: use the HaveData matcher
The test that asserts the expected NBDB state previously matched
ignoring UUIDs, because the ovsdb test framework does not provide
any sort of weak reference cleanup - as such, it is impossible for
it to remove the references of load balancers that were deleted from
logical switches / routers.

In order to make the test assert the NBDB state *with* UUIDs, we need
to patch up the UUIDs in the expected state - since the test framework
is unable to correlate the name of the load balancer to its expected
UUID once the load balancer is deleted.

This occurs because the production code only dissociates the load
balancers from the node's logical switches / GW routers for load
balancers that **are not** going to be deleted.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>