CI: Test multizone interconnect.
3 existing shard-confermance jobs and 3 existing control-plane jobs
are now testd with single node zone interconnect deployment.

Also adds a new e2e control plane job 'multi-node-zones' which deploys
a 2 zone and 2 node per zone deployment (ovn-control-plane and
ovn-worker in zone - "zone-1" and ovn-worker2 and ovn-worker3 nodes
in the zone - "zone-2") and runs a simple test. For now a single test
is added which creates a server pod in one zone and a client pod in
other zone and
  - tests connectivity between the two
  - moves the client pod scheduled node to the server pod schedule node
    zone and tests the connectivity
  - moves back the client pod node to its original zone and tests the
    connectivity again.

Note:
1. This patch skips e2e control plante tests for egress
   services and multihoming layer2 and localet networks for multi zone
   setup. The PRs which adds the support for these missing features
   should unskip the e2e tests for ic.

2. node-ip-migration job for ic is failing.  The PR which fixes this
   issue should add a e2e node-ip-migration job for ic.

Signed-off-by: Numan Siddique <numans@ovn.org>